pyproject.toml
[tool.poetry]
name = "test-project2"
version = "0.1.0"
description = ""
authors = ["natal <natal@CLOWN>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pytest = "^8.2.2"
pydantic = "^2.8.2"
dynaconf = "^3.2.5"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
unicorn = "^2.0.1.post1"
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}
fastapi-pagination = "^0.12.26"
celery = "^5.4.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
